+++ equalsverifier-11/src/main/java/nl/jqno/equalsverifier/internal/reflection/ModuleHelper.java
@@ -1,0 +1,10 @@
+package nl.jqno.equalsverifier.internal.reflection;
+
+public final class ModuleHelper {
+
+    private ModuleHelper() {}
+
+    public static boolean runsOnModulePath() {
+        return ModuleLayer.boot().findModule("nl.jqno.equalsverifier").isPresent();
+    }
+}